Move videos.md to appendix
Videos.md needs to be under the appendix folder for the right-arrow navigation UI to allow clicking/tapping to the next page when on the Glossary page. The videos.md file was incorrectly placed in the top-level navigation under src, so it only worked when clicking the Videos link via the sidebar navigation. Bug flagged here: https://linera-io.slack.com/archives/C06JRLF240Z/p1714142781061829